+++ sdk/storage/storage-file/src/highlevel.node.ts
@@ -1,6 +1,7 @@
-import * as fs from "fs";
 import { TransferProgressEvent } from "@azure/ms-rest-js";
+import * as fs from "fs";
 import { Readable } from "stream";
+
 import { Aborter } from "./Aborter";
 import { FileURL } from "./FileURL";
 import { IDownloadFromAzureFileOptions, IUploadToAzureFileOptions } from "./highlevel.common";
@@ -186,8 +187,8 @@ export async function downloadAzureFileToBuffer(
   const batch = new Batch(options.parallelism);
   for (let off = offset; off < offset + count; off = off + options.rangeSize) {
     batch.addOperation(async () => {
-      const chunkEnd = off + options.rangeSize! < count! ? off + options.rangeSize! : count!;
-      const response = await fileURL.download(aborter, off, chunkEnd - off + 1, {
+      const chunkEnd = off + options.rangeSize! <= offset + count! ? off + options.rangeSize! : offset + count!;
+      const response = await fileURL.download(aborter, off, chunkEnd - off, {
         maxRetryRequests: options.maxRetryRequestsPerRange
       });
       const stream = response.readableStreamBody!;